pyproject.toml
[project]
name = "zitadel-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
